requirements.txt
streamlit
opencv-python-headless
easyocr
numpy
pillow
